+++ src/components/MypageSidebar/components/MypageHeader.tsx
@@ -1,0 +1,10 @@
+import MypageSideList from './MypageSideList';
+import './MypageHeader.scss';
+
+export const MypageHeader = () => {
+  return (
+    <div className="mypage-header">
+      <MypageSideList />
+    </div>
+  );
+};
