+++ src/components/MypageSidebar/components/MypageSideList.tsx
@@ -1,39 +1,62 @@
-import { useEffect, useState } from 'react';
-import { Link, useLocation } from 'react-router-dom';
-import { mypagesideProps } from '../types';
+import { MyPageSideListItem } from './MypageSideListItem';
+import '../style.scss';
 
-export const MyPageSideList = ({ image, clickImage, link, text }: mypagesideProps) => {
-  const [isIcon, setIsIcon] = useState(image);
-  const location = useLocation();
+export default function MypageSideList() {
+  const MYPAGE_BASIC_URL = 'mypage';
+  const MYPAGE_BASIC_ICON = '/assets/icons';
 
-  const handleMouseEnter = () => {
-    setIsIcon(clickImage);
-  };
+  // 내 정보
+  const mypageLink = MYPAGE_BASIC_URL;
+  const mypageIcon = `${MYPAGE_BASIC_ICON}/icon-myinfo.svg`;
+  const mypageIconHover = `${MYPAGE_BASIC_ICON}/icon-myinfo-hover.svg`;
 
-  const handleMouseLeave = () => {
-    if (`/${link}` !== location.pathname) {
-      setIsIcon(image);
-    }
-  };
+  // 예약 내역
+  const reservationLink = `${MYPAGE_BASIC_URL}/reservation`;
+  const reservationIcon = `${MYPAGE_BASIC_ICON}/icon-list.svg`;
+  const reservationIconHover = `${MYPAGE_BASIC_ICON}/icon-list-hover.svg`;
 
-  useEffect(() => {
-    if (`/${link}` === location.pathname) {
-      setIsIcon(clickImage);
-    } else {
-      setIsIcon(image);
-    }
-  }, [location.pathname]);
+  // 내 공간 관리
+  const myspageAdminLink = `${MYPAGE_BASIC_URL}/admin`;
+  // const myspageAdminIcon = `${MYPAGE_BASIC_ICON}/icon-setting.svg`;
+  const myspageAdminIconHover = `${MYPAGE_BASIC_ICON}/icon-setting-hover.svg`;
+
+  // 내 공간 등록
+  const myspageAddLink = `${MYPAGE_BASIC_URL}/admin/add`;
+  // const myspageAddIcon = `${MYPAGE_BASIC_ICON}/icon-setting.svg`;
+  const myspageAddIconHover = `${MYPAGE_BASIC_ICON}/icon-setting-hover.svg`;
+
+  // 예약 현황
+  const reservationStatusLink = `${MYPAGE_BASIC_URL}/admin/status`;
+  const reservationStatusIcon = `${MYPAGE_BASIC_ICON}/icon-relstatus.svg`;
+  const reservationStatusIconHover = `${MYPAGE_BASIC_ICON}/icon-relstatus-hover.svg`;
 
   return (
-    <Link to={link}>
-      <div
-        className={`/${link}` === location.pathname ? 'mypage-side-item-current' : 'mypage-side-item'}
-        onMouseEnter={handleMouseEnter}
-        onMouseLeave={handleMouseLeave}
-      >
-        <div className="mypage-side-item-image">{isIcon && <img src={isIcon} alt={text} />}</div>
-        <div className="mypage-side-item-text">{text}</div>
-      </div>
-    </Link>
+    <div className="mapage-side-list">
+      <MyPageSideListItem text={'내정보'} link={mypageLink} image={mypageIcon} clickImage={mypageIconHover} />
+      <MyPageSideListItem
+        text={'예약 내역'}
+        link={reservationLink}
+        image={reservationIcon}
+        clickImage={reservationIconHover}
+      />
+      <MyPageSideListItem
+        text={'예약 현황'}
+        link={reservationStatusLink}
+        image={reservationStatusIcon}
+        clickImage={reservationStatusIconHover}
+      />
+      <MyPageSideListItem
+        text={'내 공간 관리'}
+        link={myspageAdminLink}
+        image={reservationIcon}
+        clickImage={myspageAdminIconHover}
+      />
+      <MyPageSideListItem
+        text={'내 공간 등록'}
+        link={myspageAddLink}
+        image={reservationIcon}
+        clickImage={myspageAddIconHover}
+      />
+    </div>
   );
-};
+}
